implement zpipeline from function
