+++ streams/shared/src/main/scala/zio/stream/ZStream.scala
@@ -146,7 +146,7 @@ final class ZStream[-R, +E, +A] private (val channel: ZChannel[R, Any, Any, Any,
   def aggregateAsync[R1 <: R, E1 >: E, A1 >: A, B](
     sink: => ZSink[R1, E1, A1, A1, B]
   )(implicit trace: Trace): ZStream[R1, E1, B] =
-    self >>> ZPipeline.aggregateAsync(sink)
+    aggregateAsyncWithin(sink, Schedule.forever)
 
   /**
    * Like `aggregateAsyncWithinEither`, but only returns the `Right` results.
@@ -162,7 +162,7 @@ final class ZStream[-R, +E, +A] private (val channel: ZChannel[R, Any, Any, Any,
     sink: => ZSink[R1, E1, A1, A1, B],
     schedule: => Schedule[R1, Option[B], Any]
   )(implicit trace: Trace): ZStream[R1, E1, B] =
-    self >>> ZPipeline.aggregateAsyncWithin(sink, schedule)
+    aggregateAsyncWithinEither(sink, schedule).collectRight
 
   /**
    * Aggregates elements using the provided sink until it completes, or until
@@ -186,8 +186,137 @@ final class ZStream[-R, +E, +A] private (val channel: ZChannel[R, Any, Any, Any,
   def aggregateAsyncWithinEither[R1 <: R, E1 >: E, A1 >: A, B, C](
     sink: => ZSink[R1, E1, A1, A1, B],
     schedule: => Schedule[R1, Option[B], C]
-  )(implicit trace: Trace): ZStream[R1, E1, Either[C, B]] =
-    self >>> ZPipeline.aggregateAsyncWithinEither(sink, schedule)
+  )(implicit trace: Trace): ZStream[R1, E1, Either[C, B]] = {
+     type HandoffSignal = ZStream.HandoffSignal[E1, A]
+     import ZStream.HandoffSignal._
+     type SinkEndReason = ZStream.SinkEndReason
+     import ZStream.SinkEndReason._
+
+     val layer =
+       ZStream.Handoff.make[HandoffSignal] <*>
+         Ref.make[SinkEndReason](ScheduleEnd) <*>
+         Ref.make(Chunk[A1]()) <*>
+         schedule.driver <*>
+         Ref.make(false) <*>
+         Ref.make(false)
+
+     ZStream.fromZIO(layer).flatMap {
+       case (handoff, sinkEndReason, sinkLeftovers, scheduleDriver, consumed, endAfterEmit) =>
+         lazy val handoffProducer: ZChannel[Any, E1, Chunk[A], Any, Nothing, Nothing, Any] =
+           ZChannel.readWithCause(
+             (in: Chunk[A]) => ZChannel.fromZIO(handoff.offer(Emit(in)).when(in.nonEmpty)) *> handoffProducer,
+             (cause: Cause[E1]) => ZChannel.fromZIO(handoff.offer(Halt(cause))),
+             (_: Any) => ZChannel.fromZIO(handoff.offer(End(UpstreamEnd)))
+           )
+
+         lazy val handoffConsumer: ZChannel[Any, Any, Any, Any, E1, Chunk[A1], Unit] =
+           ZChannel.unwrap(
+             sinkLeftovers.getAndSet(Chunk.empty).flatMap { leftovers =>
+               if (leftovers.nonEmpty) {
+                 consumed.set(true) *> ZIO.succeed(ZChannel.write(leftovers) *> handoffConsumer)
+               } else
+                 handoff.take.map {
+                   case Emit(chunk) =>
+                     ZChannel.fromZIO(consumed.set(true)) *>
+                       ZChannel.write(chunk) *>
+                       ZChannel.fromZIO(endAfterEmit.get).flatMap {
+                         case false => handoffConsumer
+                         case true  => ZChannel.unit
+                       }
+                   case Halt(cause) => ZChannel.failCause(cause)
+                   case End(ScheduleEnd) =>
+                     ZChannel.unwrap {
+                       consumed.get.map { p =>
+                         if (p) ZChannel.fromZIO(sinkEndReason.set(ScheduleEnd) *> endAfterEmit.set(true))
+                         else
+                           ZChannel
+                             .fromZIO(sinkEndReason.set(ScheduleEnd) *> endAfterEmit.set(true)) *> handoffConsumer
+                       }
+                     }
+                   case End(reason) => ZChannel.fromZIO(sinkEndReason.set(reason) *> endAfterEmit.set(true))
+                 }
+             }
+           )
+
+         def timeout(lastB: Option[B]): ZIO[R1, None.type, C] =
+           scheduleDriver.next(lastB)
+
+         def scheduledAggregator(
+           sinkFiber: Fiber.Runtime[E1, (Chunk[Chunk[A1]], B)],
+           scheduleFiber: Fiber.Runtime[None.type, C],
+           scope: Scope
+         ): ZChannel[R1, Any, Any, Any, E1, Chunk[Either[C, B]], Any] = {
+
+           val forkSink =
+             consumed.set(false) *> endAfterEmit
+               .set(false) *> (handoffConsumer pipeToOrFail sink.channel).collectElements.run.forkIn(scope)
+
+           def handleSide(leftovers: Chunk[Chunk[A1]], b: B, c: Option[C]) =
+             ZChannel.unwrap(
+               sinkLeftovers.set(leftovers.flatten) *>
+                 sinkEndReason.get.map {
+                   case ScheduleEnd =>
+                     ZChannel.unwrap {
+                       for {
+                         consumed      <- consumed.get
+                         sinkFiber     <- forkSink
+                         scheduleFiber <- timeout(Some(b)).forkIn(scope)
+                         toWrite        = c.fold[Chunk[Either[C, B]]](Chunk(Right(b)))(c => Chunk(Right(b), Left(c)))
+                       } yield
+                         if (consumed)
+                           ZChannel
+                             .write(toWrite) *> scheduledAggregator(sinkFiber, scheduleFiber, scope)
+                         else scheduledAggregator(sinkFiber, scheduleFiber, scope)
+                     }
+                   case UpstreamEnd =>
+                     ZChannel.unwrap {
+                       consumed.get.map { p =>
+                         if (p) ZChannel.write(Chunk(Right(b)))
+                         else ZChannel.unit
+                       }
+                     }
+                 }
+             )
+
+           ZChannel.unwrap {
+             sinkFiber.join.raceWith(scheduleFiber.join)(
+               (sinkExit, _) =>
+                 scheduleFiber.interrupt *>
+                   ZIO.done(sinkExit).map { case (leftovers, b) => handleSide(leftovers, b, None) },
+               (scheduleExit, _) =>
+                 ZIO
+                   .done(scheduleExit)
+                   .foldCauseZIO(
+                     _.failureOrCause match {
+                       case Left(_) =>
+                         handoff.offer(End(ScheduleEnd)).forkDaemon *> sinkFiber.join.map { case (leftovers, b) =>
+                           handleSide(leftovers, b, None)
+                         }
+                       case Right(cause) =>
+                         handoff.offer(Halt(cause)).forkDaemon *> sinkFiber.join.map { case (leftovers, b) =>
+                           handleSide(leftovers, b, None)
+                         }
+                     },
+                     c =>
+                       handoff.offer(End(ScheduleEnd)).forkDaemon *>
+                         sinkFiber.join.map { case (leftovers, b) =>
+                           handleSide(leftovers, b, Some(c))
+                         }
+                   )
+             )
+           }
+         }
+
+         ZStream.unwrapScoped[R1] {
+           for {
+             _             <- (self.channel >>> handoffProducer).run.forkScoped
+             sinkFiber     <- (handoffConsumer pipeToOrFail sink.channel).collectElements.run.forkScoped
+             scheduleFiber <- timeout(None).forkScoped
+             scope         <- ZIO.scope
+           } yield new ZStream(scheduledAggregator(sinkFiber, scheduleFiber, scope))
+         }
+     }
+   }
 
   /**
    * Maps the success values of this stream to the specified constant value.
@@ -578,8 +707,88 @@ final class ZStream[-R, +E, +A] private (val channel: ZChannel[R, Any, Any, Any,
    *   A search engine may only want to initiate a search after a user has
    *   paused typing so as to not prematurely recommend results.
    */
-  def debounce(d: => Duration)(implicit trace: Trace): ZStream[R, E, A] =
-    self >>> ZPipeline.debounce(d)
+  def debounce(d: => Duration)(implicit trace: Trace): ZStream[R, E, A] = {
+     import DebounceState._
+     import HandoffSignal._
+
+     ZStream.unwrap(
+       ZIO.transplant { grafter =>
+         for {
+           d       <- ZIO.succeed(d)
+           handoff <- ZStream.Handoff.make[HandoffSignal[E, A]]
+         } yield {
+           def enqueue(last: Chunk[A]) =
+             for {
+               f <- grafter(Clock.sleep(d).as(last).fork)
+             } yield consumer(Previous(f))
+
+           lazy val producer: ZChannel[R, E, Chunk[A], Any, E, Nothing, Any] =
+             ZChannel.readWithCause(
+               (in: Chunk[A]) =>
+                 in.lastOption.fold(producer) { last =>
+                   ZChannel.fromZIO(handoff.offer(Emit(Chunk.single(last)))) *> producer
+                 },
+               (cause: Cause[E]) => ZChannel.fromZIO(handoff.offer(Halt(cause))),
+               (_: Any) => ZChannel.fromZIO(handoff.offer(End(ZStream.SinkEndReason.UpstreamEnd)))
+             )
+
+           def consumer(state: DebounceState[E, A]): ZChannel[R, Any, Any, Any, E, Chunk[A], Any] =
+             ZChannel.unwrap(
+               state match {
+                 case NotStarted =>
+                   handoff.take.map {
+                     case Emit(last) =>
+                       ZChannel.unwrap(enqueue(last))
+                     case HandoffSignal.Halt(error) =>
+                       ZChannel.failCause(error)
+                     case HandoffSignal.End(_) =>
+                       ZChannel.unit
+                   }
+                 case Current(fiber) =>
+                   fiber.join.map {
+                     case HandoffSignal.Emit(last)  => ZChannel.unwrap(enqueue(last))
+                     case HandoffSignal.Halt(error) => ZChannel.failCause(error)
+                     case HandoffSignal.End(_)      => ZChannel.unit
+                   }
+                 case Previous(fiber) =>
+                   fiber.join
+                     .raceWith[R, E, E, HandoffSignal[E, A], ZChannel[
+                       R,
+                       Any,
+                       Any,
+                       Any,
+                       E,
+                       Chunk[A],
+                       Any
+                     ]](
+                       handoff.take
+                     )(
+                       {
+                         case (Exit.Success(a), current) =>
+                           ZIO.succeedNow(ZChannel.write(a) *> consumer(Current(current)))
+                         case (Exit.Failure(cause), current) =>
+                           current.interrupt as ZChannel.failCause(cause)
+                       },
+                       {
+                         case (Exit.Success(Emit(last)), previous) =>
+                           previous.interrupt *> enqueue(last)
+                         case (Exit.Success(Halt(cause)), previous) =>
+                           previous.interrupt as ZChannel.failCause(cause)
+                         case (Exit.Success(End(_)), previous) =>
+                           previous.join.map(ZChannel.write(_) *> ZChannel.unit)
+                         case (Exit.Failure(cause), previous) =>
+                           previous.interrupt as ZChannel.failCause(cause)
+                       }
+                     )
+               }
+             )
+
+           ZStream.scoped[R]((self.channel >>> producer).runScoped.forkScoped) *>
+             new ZStream(consumer(NotStarted))
+         }
+       }
+     )
+   }
 
   /**
    * Taps the stream, printing the result of calling `.toString` on the emitted
